GripUp: Split files
